implement user accessibility to characters created

+++ EFPractice/Services/CharacterService.cs
@@ -2,10 +2,12 @@
 using EFPractice.Data;
 using EFPractice.Dtos;
 using EFPractice.Models;
+using Microsoft.AspNetCore.Http;
 using Microsoft.EntityFrameworkCore;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
 
 namespace EFPractice.Services
@@ -15,19 +17,27 @@ namespace EFPractice.Services
 
         private readonly IMapper _mapper;
         private readonly DataContext _context;
+        private readonly IHttpContextAccessor _httpContextAccessor;
 
-        public CharacterService(IMapper mapper, DataContext context)
+        public CharacterService(IMapper mapper, DataContext context, IHttpContextAccessor httpContextAccessor )
         {
             _mapper = mapper;
             _context = context;
+            _httpContextAccessor = httpContextAccessor;
         }
+
+        private int GetUserId() => int.Parse(_httpContextAccessor.HttpContext.User.FindFirstValue(ClaimTypes.NameIdentifier));
+
+
         public async Task<ServiceResponse<List<GetCharacterDto>>> AddCharacter(AddCharacterDto newCharacter)
         {   
             ServiceResponse<List<GetCharacterDto>> serviceresponse = new ServiceResponse<List<GetCharacterDto>>();
             Character character = _mapper.Map<Character>(newCharacter);
+
+            character.User = await _context.Users.FirstOrDefaultAsync(u => u.Id == GetUserId());
             await _context.Characters.AddAsync(character );
             await _context.SaveChangesAsync();
-            serviceresponse.Data = _context.Characters.Select(c => _mapper.Map<GetCharacterDto>(c)).ToList();
+            serviceresponse.Data = _context.Characters.Where(u => u.User.Id == GetUserId()).Select(c => _mapper.Map<GetCharacterDto>(c)).ToList();
             return serviceresponse;
             
         }
@@ -35,7 +45,7 @@ namespace EFPractice.Services
         public async Task<ServiceResponse<List<GetCharacterDto>>> GetAllCharacters()
         {
             ServiceResponse<List<GetCharacterDto>> serviceresponse = new ServiceResponse<List<GetCharacterDto>>();
-            List<Character> dbCharacters = await _context.Characters.ToListAsync();
+            List<Character> dbCharacters = await _context.Characters.Where(c => c.User.Id == GetUserId()).ToListAsync();
             if (dbCharacters == null)
             {
                 serviceresponse.Success = false;
@@ -48,7 +58,9 @@ namespace EFPractice.Services
         public async Task<ServiceResponse<GetCharacterDto>> GetCharacterById(int id)
         {
             ServiceResponse<GetCharacterDto> serviceresponse = new ServiceResponse<GetCharacterDto>();
-            Character dbcharacter = await _context.Characters.FirstOrDefaultAsync(x => x.Id == id);
+            Character dbcharacter = await _context.Characters
+                        .FirstOrDefaultAsync(x => x.Id == id && x.User.Id == GetUserId());
+            
             if (dbcharacter == null)
             {
                 serviceresponse.Success = false;
@@ -64,12 +76,15 @@ namespace EFPractice.Services
             try
             {
 
-                Character character = await _context.Characters.FirstOrDefaultAsync(c => c.Id == newCHar.Id);
+                Character character = await _context.Characters
+                    .FirstOrDefaultAsync(c => c.Id == newCHar.Id && c.User.Id == GetUserId());
                 if (character == null)
                 {
                     serviceResponse.Success = false;
+                    serviceResponse.Message = "Character not found";
                     return serviceResponse;
                 }
+
                 character.Name = newCHar.Name;
                 character.Intelligence = newCHar.Intelligence;
                 character.Strength = newCHar.Strength;
@@ -100,10 +115,19 @@ namespace EFPractice.Services
             try
             {
 
-                Character character = _context.Characters.First(c => c.Id == id);
-                _context.Characters.Remove(character);
-                await _context.SaveChangesAsync();
-                serviceResponse.Data = _context.Characters.Select(c => _mapper.Map<GetCharacterDto>(c)).ToList();
+                Character character =await _context.Characters.FirstOrDefaultAsync(c => c.Id == id && c.User.Id == GetUserId());
+                if (character != null)
+                {
+                    _context.Characters.Remove(character);
+                    await _context.SaveChangesAsync();
+                    serviceResponse.Data = _context.Characters.Where(x => x.User.Id == GetUserId()).Select(c => _mapper.Map<GetCharacterDto>(c)).ToList();
+                }
+                else
+                {
+                    serviceResponse.Success = false;
+                    serviceResponse.Message = "Character doesn't exist";
+                    return serviceResponse;
+                }
             }
             catch (Exception e)
             {
